- Correção dos campos valores de double para BigDecimal

+++ src/main/kotlin/br/com/alura/alugames/model/InfoJogoJson.kt
@@ -1,7 +1,9 @@
 package br.com.alura.alugames.model
 
+import java.math.BigDecimal
+
 data class InfoJogoJson(
     val titulo: String,
     val capa: String,
-    val preco: Double,
+    val preco: BigDecimal,
     val descricao: String)
