+++ src/main/kotlin/br/com/alura/alugames/model/Plano.kt
@@ -1,7 +1,9 @@
 package br.com.alura.alugames.model
 
+import java.math.BigDecimal
+
 sealed class Plano(val tipo:String) {
-    open fun obterValorAluguel(aluguel:Aluguel):Double{
-        return  aluguel.jogo.preco * aluguel.periodo.emDias
+    open fun obterValorAluguel(aluguel:Aluguel): BigDecimal {
+        return  aluguel.jogo.preco * BigDecimal(aluguel.periodo.emDias)
     }
 }
